+++ src/OpcUaStackCore/BuildInTypes/OpcUaNodeIdBase.cpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2015-2018 Kai Huebl (kai@huebl-sgh.de
+   Copyright 2015-2019 Kai Huebl (kai@huebl-sgh.de
 
    Lizenziert gemäß Apache Licence Version 2.0 (die „Lizenz“); Nutzung dieser
    Datei nur in Übereinstimmung mit der Lizenz erlaubt.
@@ -226,9 +226,9 @@ namespace OpcUaStackCore
 	}
 
 	void
-	OpcUaNodeIdBase::copyFrom(OpcUaNodeIdBase& opcUaNodeIdBase)
+	OpcUaNodeIdBase::copyFrom(const OpcUaNodeIdBase& opcUaNodeIdBase)
 	{
-		opcUaNodeIdBase.copyTo(*this);
+		const_cast<OpcUaNodeIdBase*>(&opcUaNodeIdBase)->copyTo(*this);
 	}
 
 	void