+++ src/OpcUaStackServer/StandardObjectType/build.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-OpcUaObjectTypeGenerator4 --nodeset Opc.Ua.NodeSet.xml --variabletype BaseObjectType --buildSubTypes 1
+NODESETS="--nodeset Opc.Ua.NodeSet2.xml --nodeset Opc.Ua.NodeSet2.Part14.xml --nodeset Opc.Ua.NodeSet2.Services.xml"
+
+OpcUaObjectTypeGenerator4 ${NODESETS} --objecttype BaseObjectType --buildSubTypes 1