+++ src/OpcUaStackServer/StandardVariableType/build.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-OpcUaVariableTypeGenerator4 --nodeset Opc.Ua.NodeSet.xml --variabletype BaseVariableType --buildSubTypes 1
+NODESETS="--nodeset Opc.Ua.NodeSet2.xml --nodeset Opc.Ua.NodeSet2.Part14.xml --nodeset Opc.Ua.NodeSet2.Services.xml"
+
+OpcUaVariableTypeGenerator4 ${NODESETS} --variabletype BaseVariableType --buildSubTypes 1